Move stacktrace normalization into ExceptionNormalizer
