+++ src/core/exception/index.ts
@@ -1,1 +1,3 @@
+export * from './is-src-frame';
+export * from './normalize-frame-path';
 export * from './jest-skip-source-file';
