+++ src/core/exception/is-src-frame.ts
@@ -1,0 +1,7 @@
+export const isSrcFrame = (frame: string) =>
+  frame.startsWith('    at')
+    ? !frame.includes('node_modules') &&
+      !frame.includes('(internal/') &&
+      !frame.includes('(node:') &&
+      !frame.includes('(<anonymous>)')
+    : true;
